requirements.txt
praw==7.2.0
Click==8.0.0
pyyaml==5.4.1
